module-requirement.txt
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.4
django-classy-tags==0.9.0
django-cms==3.6.0
django-debug-toolbar==2.0
django-filer==1.5.0
django-filter==2.1.0
django-formtools==2.1
django-haystack==2.8.1
django-js-asset==1.2.2
django-mptt==0.10.0
django-polymorphic==2.0.3
django-sekizai==1.0.0
django-treebeard==4.3
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.1.0
djangocms-column==1.9.0
djangocms-file==2.3.0
djangocms-googlemap==1.3.0
djangocms-installer==1.1.0
djangocms-link==2.4.0
djangocms-picture==2.0.8
djangocms-snippet==2.2.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.8.0
djangocms-video==2.1.1
djangorestframework==3.9.4
easy-thumbnails==2.6
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
Markdown==3.1.1
mysql==0.0.2
mysql-connector-python==8.0.17
mysqlclient==1.4.2.post1
oauthlib==3.1.0
pbr==5.2.1
Pillow==6.0.0
pkg-resources==0.0.0
protobuf==3.9.1
PySocks==1.7.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
stevedore==1.30.1
tweepy==3.8.0
tzlocal==1.5.1
Unidecode==1.0.23
urllib3==1.25.3
virtualenv==16.6.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
webencodings==0.5.1
